backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
motor==3.6.0
pymongo==4.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.20
httpx==0.28.1
openai==1.59.7
pytest==8.3.4
redis==5.0.7
numpy==1.26.4
tensorflow==2.16.1
keras==3.13.1
pypdf==5.1.0
python-docx==1.1.2
psutil==7.0.0
apscheduler==3.10.4
